+++ src/main/java/com/xhadl/yournotion/Repository/SurveyParticipantRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SurveyParticipantRepository extends JpaRepository<SurveyParticipantEntity, Long> {
+public interface SurveyParticipantRepository extends JpaRepository<SurveyParticipantEntity, Integer> {
     public SurveyParticipantEntity save(SurveyAvailableEntity surveyAvailableEntity);
     public int countBySurveyId(int survey_id);
     public SurveyParticipantEntity findBySurveyIdAndParticipantId(int surveyId, int participantId);
